Fix output format
*Issue #, if available:*

None

*Description of changes:*

Jimp's getBuffer returns the contents in the original format, in this case it's PNG.
In order to generate the JPEG file, we temporarily save it to /tmp/, and then upload the converted file.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.